857: Remove legacy calendar export code
