--- app/models/calendars/exports/all_meals_export.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-module Calendars
-  module Exports
-    # Exports all meals in cluster
-    class AllMealsExport < MealsExport
-      protected
-
-      def scope
-        base_scope
-      end
-    end
-  end
-end
