--- app/models/calendars/exports/type_error.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Calendars
-  module Exports
-    class TypeError < StandardError
-    end
-  end
-end
