--- spec/models/calendars/exports/legacy_ical_generator_spec.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-require "rails_helper"
-
-describe Calendars::Exports::LegacyIcalGenerator do
-  let(:export) do
-    double(calendar_name: "Foo Calendar", events: events, kind_name: "Thing")
-  end
-  subject(:ical) { described_class.new(export).generate }
-
-  context "with description array inluding nils" do
-    let(:events) { [build(:calendar_export_event, description: ["Short", nil, "Fluff"])] }
-
-    it "breaks lines properly and adds url" do
-      is_expected.to match(%r{^DESCRIPTION:Short\s+\n Fluff\s+\n https://you.gather.coop/stuff/123\s+$})
-    end
-  end
-end
